refactor(estimation): extract fixed-eta observation-NLL gradients from saem
`obs_nll_subject_grad`, `obs_nll_subject_grad_iov` and the
`obs_nll_subject_into_iov` evaluator they depend on are the per-subject
gradient of the observation NLL with the random effects held *fixed* --
deliberately without FOCE's `∂η̂/∂θ` implicit-function terms or its
`½log|H̃|` curvature. That makes them the right primitive for any estimator
whose η is given rather than profiled, which is SAEM's M-step today and
variational inference next.

Moved verbatim into `estimation/fixed_eta_gradient.rs` with their FD-parity
tests in a sibling `_tests.rs`. The only changes to the moved production code
are three `fn` -> `pub(crate) fn`; the only changes to the moved tests are
rustfmt reflowing two call sites that gained horizontal room after
de-indenting. Verified by diffing both blocks against the pre-move file.
`obs_nll_sum` becomes `pub(crate)` so the tests can still reach their FD
reference.

Re-extracted against current `main` rather than the branch's original base.
`main` has since landed #484/#1029 (apply the residual magnitude in every
estimator), which rewrote all three of these functions -- the `_scaled`
variance/derivative forms and the `gaussian_obs_nll_scaled` theta-FD branch
that carries the direct theta -> variance channel. Extracting the pre-#1029
snapshot would have compiled cleanly while silently reverting the weighted
error model for SAEM. The two FD-parity tests #1029 added
(`obs_nll_subject_grad_weighted_error_matches_fd` and
`obs_nll_subject_grad_theta_dependent_magnitude_matches_fd`, plus their shared
`check_saem_mstep_matches_fd` body) move alongside the code, so the sibling
now holds all six.

No behaviour change: SAEM is byte-identical by construction, not just by test.
