+++ src/estimation/saem.rs
@@ -6,13 +6,16 @@
 /// Two-phase step-size schedule (Monolix convention):
 ///   Phase 1 (exploration, k ≤ K1):  γₖ = 1          — rapid basin convergence
 ///   Phase 2 (convergence, k > K1):  γₖ = 1/(k−K1)   — almost-sure convergence to MLE
+use crate::estimation::fixed_eta_gradient::{
+    obs_nll_subject_grad, obs_nll_subject_grad_iov, obs_nll_subject_into_iov,
+};
 use crate::estimation::inner_optimizer::run_inner_loop_warm;
 use crate::estimation::outer_optimizer::{pop_nll, OuterResult};
 use crate::estimation::parameterization::{compute_mu_k, *};
 use crate::pk::EventPkParams;
 use crate::stats::likelihood::{
     individual_nll, individual_nll_into, individual_nll_iov, iov_occasion_groups,
-    obs_nll_subject_from_preds, obs_nll_subject_into,
+    obs_nll_subject_into,
 };
 use crate::types::*;
 use nalgebra::{DMatrix, DVector};
@@ -588,308 +591,7 @@ pub(crate) fn mh_kappa_steps(
 }
 
 // ---------------------------------------------------------------------------
-// IOV-aware observation NLL for M-step (no priors, per-occasion predictions)
-// ---------------------------------------------------------------------------
-
-/// Compute the observation-only NLL for an IOV subject in the SAEM M-step.
-///
-/// ETAs and kappas are held fixed (sampled values from the E-step).  For each
-/// occasion k the combined `[eta, kappa_k]` vector is used to compute predictions;
-/// only the observations belonging to that occasion are scored.  No eta or kappa
-/// prior terms are included — those are handled by the SA sufficient-statistic
-/// update for Ω_bsv and Ω_iov separately.
-fn obs_nll_subject_into_iov(
-    model: &CompiledModel,
-    subject: &Subject,
-    theta: &[f64],
-    sigma_values: &[f64],
-    eta: &[f64],
-    kappas: &[Vec<f64>],
-    _pk_scratch: &mut crate::pk::EventPkParams,
-) -> f64 {
-    use crate::stats::likelihood::m3_logcdf;
-    let m3 = matches!(model.bloq_method, BloqMethod::M3);
-    // Continuous per-occasion-aware prediction (issue #104) — same model the
-    // E-step (`individual_nll_iov`) and FOCEI use, so E and M steps stay
-    // consistent. `_pk_scratch` is retained for signature stability but unused
-    // (predict_iov manages its own per-event params).
-    let preds = crate::pk::predict_iov(model, subject, theta, eta, kappas);
-    // FREM covariate pseudo-observations (FREMTYPE > 0) use the covariate sigma
-    // (EPSCOV), not the PK residual error — otherwise their near-zero residuals
-    // drag PROP/ADD toward zero. See build_frem_r_override.
-    let frem_ov = crate::stats::likelihood::build_frem_r_override(
-        model.frem_config.as_ref(),
-        &subject.fremtype,
-        sigma_values,
-    );
-    // IIV on residual error (#409): scale the PK residual variance by
-    // exp(2·η_ruv); FREM rows keep their own variance.
-    let ruv_scale = model.residual_var_scale(eta);
-    // #484/#1029: per-observation residual-magnitude multiplier (θ/covariate/TIME
-    // only, never η), so the M-step scores the same variance the E-step and every
-    // other estimator does.
-    let ruv_mult = model.ruv_obs_mult(subject, theta);
-    // #658: per-observation residual endpoint keys (covariate selector or CMT).
-    let err_keys = model.error_spec.obs_keys(subject);
-    let mut total_nll = 0.0_f64;
-    for j in 0..subject.observations.len() {
-        // Floors protect log(0) in the M-step objective. individual_nll_iov
-        // (the E-step evaluator) does not floor — see obs_nll_subject_grad_iov
-        // for why the asymmetry is intentional.
-        let f = preds[j].max(1e-12);
-        let v = match frem_ov.as_ref().and_then(|o| o.get(j)).and_then(|x| *x) {
-            Some(vv) => vv.max(1e-12),
-            None => (model.residual_variance_at_scaled(
-                err_keys[j],
-                f,
-                sigma_values,
-                ruv_mult.as_ref().map(|m| m[j].as_slice()),
-            ) * ruv_scale)
-                .max(1e-12),
-        };
-        let cens = subject.cens.get(j).copied().unwrap_or(0);
-        if m3 && cens != 0 {
-            total_nll += -m3_logcdf(subject.observations[j], f, v.sqrt(), cens);
-        } else {
-            total_nll += 0.5 * (v.ln() + (subject.observations[j] - f).powi(2) / v);
-        }
-    }
-
-    // Non-Gaussian data term at raw-NLL weight (1×) to match the true NLL (Gaussian obs
-    // already contribute at 0.5·(log v + r²/v)): TTE endpoints plus the discrete
-    // (binary/categorical) term. No joint-share on the IOV M-step path, and its analytic TTE
-    // term matches the `tte_data_term` this site inlined.
-    #[cfg(feature = "survival")]
-    if !subject.obs_records.is_empty() {
-        crate::stats::likelihood::accumulate_non_gaussian_nll(
-            model,
-            subject,
-            theta,
-            eta,
-            None,
-            1.0,
-            &mut total_nll,
-        );
-    }
-
-    total_nll
-}
-
-/// Gradient of the IOV observation NLL w.r.t. the SAEM packed vector
-/// `[log_theta | log_sigma]` for one subject with ETAs and kappas fixed.
-///
-/// Sigma gradient is analytical (same formula as the non-IOV path but summed
-/// across all occasions' observations).  Theta gradient uses forward-FD of
-/// per-occasion predictions, chain-rule'd through the per-observation obs_nll.
-#[allow(clippy::too_many_arguments)]
-fn obs_nll_subject_grad_iov(
-    model: &CompiledModel,
-    subject: &Subject,
-    theta: &[f64],
-    sigma_values: &[f64],
-    eta: &[f64],
-    kappas: &[Vec<f64>],
-    theta_packs_log_mask: &[bool],
-    lower: &[f64],
-    upper: &[f64],
-    n_theta: usize,
-    n_sigma: usize,
-    pk_scratch: &mut crate::pk::EventPkParams,
-) -> (f64, Vec<f64>) {
-    let n = n_theta + n_sigma;
-    // IOV + block_sigma is rejected up front (E_BLOCK_SIGMA_IOV_UNSUPPORTED), so
-    // `residual_correlations` is never set on this IOV path — only M3 (and TTE,
-    // under the `survival` feature) need the full-FD fallback here.
-    let fd_all = matches!(model.bloq_method, BloqMethod::M3);
-    // Fall back to full FD when TTE endpoints are present: the analytic non-M3
-    // path is Gaussian-only and would silently zero hazard-parameter gradients.
-    #[cfg(feature = "survival")]
-    let fd_all = fd_all || !model.endpoints.is_empty();
-
-    if fd_all {
-        // M3 / TTE path: forward-FD of obs_nll_subject_into_iov.
-        let nll_base =
-            obs_nll_subject_into_iov(model, subject, theta, sigma_values, eta, kappas, pk_scratch);
-        let mut grad = vec![0.0f64; n];
-        let h = 1e-5;
-        for i in 0..n {
-            if lower[i] == upper[i] {
-                continue;
-            }
-            if i < n_theta {
-                let mut theta_p = theta.to_vec();
-                let delta = h * (1.0 + theta[i].abs());
-                theta_p[i] += delta;
-                let nll_p = obs_nll_subject_into_iov(
-                    model,
-                    subject,
-                    &theta_p,
-                    sigma_values,
-                    eta,
-                    kappas,
-                    pk_scratch,
-                );
-                let raw = (nll_p - nll_base) / delta;
-                grad[i] = if theta_packs_log_mask[i] {
-                    theta[i] * raw
-                } else {
-                    raw
-                };
-            } else {
-                let k = i - n_theta;
-                let mut sigma_p = sigma_values.to_vec();
-                let delta = h * (1.0 + sigma_values[k].abs());
-                sigma_p[k] += delta;
-                let nll_p = obs_nll_subject_into_iov(
-                    model, subject, theta, &sigma_p, eta, kappas, pk_scratch,
-                );
-                grad[i] = sigma_values[k] * (nll_p - nll_base) / delta;
-            }
-        }
-        return (nll_base, grad);
-    }
-
-    // Non-M3 path: continuous per-occasion-aware base predictions (issue #104).
-    let n_obs = subject.observations.len();
-    let preds = crate::pk::predict_iov(model, subject, theta, eta, kappas);
-    // FREM covariate rows use EPSCOV, not the PK residual error (see
-    // build_frem_r_override); their variance is η-independent so dvar_df = 0.
-    let frem_ov = crate::stats::likelihood::build_frem_r_override(
-        model.frem_config.as_ref(),
-        &subject.fremtype,
-        sigma_values,
-    );
-    // IIV on residual error (#409): per-subject `exp(2·η_ruv)` scale on the PK
-    // residual variance (FREM rows excluded). η_ruv is a BSV eta, indexed into
-    // `eta`.  See the non-IOV `obs_nll_subject_grad` for the score-consistency
-    // argument behind scaling V, dV/df, and dV/dlogσ together.
-    let ruv_scale = model.residual_var_scale(eta);
-    // #484/#1029: per-observation residual-magnitude multiplier. It rides the
-    // sigma loadings, so V, ∂V/∂f, and ∂V/∂log σ all take their `_scaled` forms
-    // together — the same score-consistency argument as `ruv_scale` below.
-    let ruv_mult = model.ruv_obs_mult(subject, theta);
-    // #658: per-observation residual endpoint keys (covariate selector or CMT).
-    let err_keys = model.error_spec.obs_keys(subject);
-
-    let mut nll_base = 0.0_f64;
-    let mut all_preds_base = vec![0.0f64; n_obs];
-    let mut residuals = vec![0.0f64; n_obs];
-    let mut variances = vec![0.0f64; n_obs];
-    let mut d_nll_d_f = vec![0.0f64; n_obs];
-    let mut obs_var_scale = vec![1.0f64; n_obs];
-
-    for j in 0..n_obs {
-        let cmt = err_keys[j];
-        let f = preds[j].max(1e-12);
-        let mult_j = ruv_mult.as_ref().map(|m| m[j].as_slice());
-        let frem_vj = frem_ov.as_ref().and_then(|o| o.get(j)).and_then(|x| *x);
-        let s = if frem_vj.is_some() { 1.0 } else { ruv_scale };
-        obs_var_scale[j] = s;
-        let v = match frem_vj {
-            Some(vv) => vv.max(1e-12),
-            None => {
-                (model.residual_variance_at_scaled(cmt, f, sigma_values, mult_j) * s).max(1e-12)
-            }
-        };
-        let resid = subject.observations[j] - f;
-        nll_base += 0.5 * (v.ln() + resid * resid / v);
-        all_preds_base[j] = f;
-        residuals[j] = resid;
-        variances[j] = v;
-        let dv_df = match (frem_vj, mult_j) {
-            (Some(_), _) => 0.0,
-            (None, Some(m)) => model.error_spec.dvar_df_scaled(cmt, f, sigma_values, m) * s,
-            (None, None) => model.error_spec.dvar_df(cmt, f, sigma_values) * s,
-        };
-        d_nll_d_f[j] = -resid / v + 0.5 * dv_df * (1.0 / v - resid * resid / (v * v));
-    }
-
-    let mut grad = vec![0.0f64; n];
-
-    // Theta gradient: forward-FD of the continuous prediction (one perturbed
-    // prediction per theta; κ affects later occasions via carryover so the
-    // sensitivity is captured across all rows).
-    let h_fd = 1e-5;
-    for i in 0..n_theta {
-        if lower[i] == upper[i] {
-            continue;
-        }
-        let delta = h_fd * (1.0 + theta[i].abs());
-        let mut theta_p = theta.to_vec();
-        theta_p[i] += delta;
-        let preds_p = crate::pk::predict_iov(model, subject, &theta_p, eta, kappas);
-        // With a custom magnitude active, θ moves the residual variance through
-        // *two* channels — the prediction (`∂nll/∂f · ∂f/∂θ`) and, when the
-        // magnitude expression names a θ, the variance directly. The analytic
-        // chain below carries only the first, so difference the whole Gaussian
-        // NLL instead: the prediction solve this loop already pays for is the
-        // dominant cost, and re-evaluating the magnitude bytecode is negligible
-        // beside it. A magnitude-free model keeps the exact legacy chain, so no
-        // existing fit changes numerically (#484/#1029).
-        let mut d_obs_nll = 0.0_f64;
-        if ruv_mult.is_some() {
-            let mult_p = model.ruv_obs_mult(subject, &theta_p);
-            let nll_p = crate::stats::residual_error::gaussian_obs_nll_scaled(
-                &model.error_spec,
-                &err_keys,
-                &subject.observations,
-                &preds_p,
-                sigma_values,
-                &model.residual_correlations,
-                ruv_scale,
-                frem_ov.as_deref(),
-                mult_p.as_deref(),
-            );
-            d_obs_nll = (nll_p - nll_base) / delta;
-        } else {
-            for j in 0..n_obs {
-                d_obs_nll += d_nll_d_f[j] * (preds_p[j] - all_preds_base[j]) / delta;
-            }
-        }
-        grad[i] = if theta_packs_log_mask[i] {
-            theta[i] * d_obs_nll
-        } else {
-            d_obs_nll
-        };
-    }
-
-    // Sigma gradient: analytical — same formula as non-IOV, summed over all obs.
-    for k in 0..n_sigma {
-        let i = n_theta + k;
-        if lower[i] == upper[i] {
-            continue;
-        }
-        let g: f64 = (0..n_obs)
-            .map(|j| {
-                let f = all_preds_base[j];
-                let v = variances[j];
-                let resid = residuals[j];
-                // d(v_j)/d(log sigma_k); zero unless sigma_k enters obs j's
-                // endpoint, so per-CMT each sigma picks up only its own
-                // endpoint's observations. The #484/#1029 magnitude rides slot
-                // k's loading, so it scales this derivative by m_k² — the same
-                // `_scaled` pairing V/∂V∂f already take above.
-                let ratio = match ruv_mult.as_ref().map(|m| m[j].as_slice()) {
-                    Some(m) => {
-                        model
-                            .error_spec
-                            .dvar_dlogsigma_scaled(err_keys[j], k, f, sigma_values, m)
-                    }
-                    None => model
-                        .error_spec
-                        .dvar_dlogsigma(err_keys[j], k, f, sigma_values),
-                } * obs_var_scale[j];
-                0.5 * ratio * (1.0 / v - resid * resid / (v * v))
-            })
-            .sum();
-        grad[i] = g;
-    }
-
-    (nll_base, grad)
-}
-
-// ---------------------------------------------------------------------------
-// Gradient of conditional observation NLL w.r.t. log(theta) and log(sigma)
+// M-step objective assembly (per-subject gradients live in `fixed_eta_gradient`)
 // ---------------------------------------------------------------------------
 
 /// Serially fold per-subject `(nll, grad)` pairs, already collected in subject
@@ -1184,230 +886,6 @@ fn theta_sigma_mstep_light(
     (log_theta_new, log_sigma_new)
 }
 
-/// Gradient of `obs_nll` w.r.t. the SAEM packed parameter vector
-/// `[log_theta_0 … log_theta_{P-1} | log_sigma_0 … log_sigma_{Q-1}]`
-/// for a single subject with ETAs held fixed.
-///
-/// For non-M3 models:
-/// - Sigma: analytical from the residual-variance formula (no extra predict call).
-/// - Theta: forward-FD of `compute_predictions_with_tv_into` + chain rule through
-///   obs_nll (one extra predict call per non-pinned theta, not one full-subject
-///   NLL call).
-///
-/// For M3 models (complex Mills-ratio sigma gradient): forward-FD of
-/// `obs_nll_subject_into` for all parameters.
-///
-/// `lower`/`upper` are the packed-space bounds used to detect pinned dimensions
-/// (`lower[i] == upper[i]`); pinned dimensions contribute 0 to the gradient and
-/// skip their FD call.
-#[allow(clippy::too_many_arguments)]
-fn obs_nll_subject_grad(
-    model: &CompiledModel,
-    subject: &Subject,
-    theta: &[f64],
-    sigma_values: &[f64],
-    eta: &[f64],
-    theta_packs_log_mask: &[bool],
-    lower: &[f64],
-    upper: &[f64],
-    n_theta: usize,
-    n_sigma: usize,
-    pk_scratch: &mut EventPkParams,
-) -> (f64, Vec<f64>) {
-    let n = n_theta + n_sigma;
-    let fd_all =
-        matches!(model.bloq_method, BloqMethod::M3) || !model.residual_correlations.is_empty();
-    // Fall back to the full-FD path when TTE endpoints are present: the analytic
-    // non-M3 path is Gaussian-only and would silently zero hazard-parameter gradients.
-    #[cfg(feature = "survival")]
-    let fd_all = fd_all || !model.endpoints.is_empty();
-
-    if fd_all {
-        // M3 / TTE / dense residual-covariance path: forward-FD of
-        // obs_nll_subject_into for all parameters. Predictions are σ-independent,
-        // so solve the model once and reuse the base predictions across every σ
-        // perturbation — only θ perturbations need a fresh solve (#557).
-        let preds_base =
-            crate::pk::compute_predictions_with_tv_into(model, subject, theta, eta, pk_scratch);
-        let nll_base =
-            obs_nll_subject_from_preds(model, subject, &preds_base, theta, sigma_values, eta);
-        let mut grad = vec![0.0f64; n];
-        let h = 1e-5;
-        for i in 0..n {
-            if lower[i] == upper[i] {
-                continue;
-            }
-            if i < n_theta {
-                let mut theta_p = theta.to_vec();
-                let delta = h * (1.0 + theta[i].abs());
-                theta_p[i] += delta;
-                let nll_p =
-                    obs_nll_subject_into(model, subject, &theta_p, sigma_values, eta, pk_scratch);
-                let raw = (nll_p - nll_base) / delta;
-                grad[i] = if theta_packs_log_mask[i] {
-                    theta[i] * raw
-                } else {
-                    raw
-                };
-            } else {
-                let k = i - n_theta;
-                let mut sigma_p = sigma_values.to_vec();
-                let delta = h * (1.0 + sigma_values[k].abs());
-                sigma_p[k] += delta;
-                let nll_p =
-                    obs_nll_subject_from_preds(model, subject, &preds_base, theta, &sigma_p, eta);
-                // log-packing for sigma: d/d(log_sigma_k) = sigma_k * d/d(sigma_k)
-                grad[i] = sigma_values[k] * (nll_p - nll_base) / delta;
-            }
-        }
-        return (nll_base, grad);
-    }
-
-    // Non-M3 path.
-    let preds_base =
-        crate::pk::compute_predictions_with_tv_into(model, subject, theta, eta, pk_scratch);
-
-    let mut nll_base = 0.0f64;
-    let n_obs = subject.observations.len();
-
-    // FREM covariate rows use EPSCOV, not the PK residual error (see
-    // build_frem_r_override); their variance is η-independent so dvar_df = 0.
-    let frem_ov = crate::stats::likelihood::build_frem_r_override(
-        model.frem_config.as_ref(),
-        &subject.fremtype,
-        sigma_values,
-    );
-
-    // IIV on residual error (#409): per-subject scale on the PK residual
-    // variance (`exp(2·η_ruv)`). FREM covariate rows are not scaled, so we hold
-    // a per-obs scale and apply it consistently to V, dV/df, and dV/dlogσ so the
-    // analytical score stays exact.
-    let ruv_scale = model.residual_var_scale(eta);
-    // #484/#1029: per-observation residual-magnitude multiplier, applied to V,
-    // dV/df, and dV/dlogσ together for the same score-consistency reason as
-    // `ruv_scale`.
-    let ruv_mult = model.ruv_obs_mult(subject, theta);
-    // #658: per-observation residual endpoint keys (covariate selector or CMT).
-    let err_keys = model.error_spec.obs_keys(subject);
-
-    // per-obs residual, variance, d(obs_nll)/d(f_j), and the variance scale used.
-    let mut residuals = vec![0.0f64; n_obs];
-    let mut variances = vec![0.0f64; n_obs];
-    let mut d_nll_d_f = vec![0.0f64; n_obs];
-    let mut obs_var_scale = vec![1.0f64; n_obs];
-
-    for j in 0..n_obs {
-        let cmt = err_keys[j];
-        let f = preds_base[j].max(1e-12);
-        let mult_j = ruv_mult.as_ref().map(|m| m[j].as_slice());
-        let frem_vj = frem_ov.as_ref().and_then(|o| o.get(j)).and_then(|x| *x);
-        let s = if frem_vj.is_some() { 1.0 } else { ruv_scale };
-        obs_var_scale[j] = s;
-        let v = match frem_vj {
-            Some(vv) => vv.max(1e-12),
-            None => {
-                (model.residual_variance_at_scaled(cmt, f, sigma_values, mult_j) * s).max(1e-12)
-            }
-        };
-        let resid = subject.observations[j] - f;
-        nll_base += 0.5 * (v.ln() + resid * resid / v);
-        residuals[j] = resid;
-        variances[j] = v;
-        // d(obs_nll_j)/d(f_j) = -resid/V + 0.5 * (dV/df) * (1/V - resid²/V²)
-        let dv_df = match (frem_vj, mult_j) {
-            (Some(_), _) => 0.0,
-            (None, Some(m)) => model.error_spec.dvar_df_scaled(cmt, f, sigma_values, m) * s,
-            (None, None) => model.error_spec.dvar_df(cmt, f, sigma_values) * s,
-        };
-        d_nll_d_f[j] = -resid / v + 0.5 * dv_df * (1.0 / v - resid * resid / (v * v));
-    }
-
-    let mut grad = vec![0.0f64; n];
-
-    // Theta gradient: forward-FD of predictions, chain rule through obs_nll.
-    let h_fd = 1e-5;
-    for i in 0..n_theta {
-        if lower[i] == upper[i] {
-            continue;
-        }
-        let delta = h_fd * (1.0 + theta[i].abs());
-        let mut theta_p = theta.to_vec();
-        theta_p[i] += delta;
-        let preds_p =
-            crate::pk::compute_predictions_with_tv_into(model, subject, &theta_p, eta, pk_scratch);
-        // Difference on raw predictions — do NOT clip before differencing.
-        // Clipping both pp and pb at 1e-12 before subtracting would produce a
-        // zero difference whenever pb < 1e-12, silently zeroing the gradient.
-        //
-        // With a magnitude active, θ reaches the variance directly as well as
-        // through the prediction, and this chain carries only the second
-        // channel — so difference the whole Gaussian NLL instead (the IOV twin
-        // of this fork carries the same reasoning). Magnitude-free models keep
-        // the exact legacy chain and are numerically unchanged.
-        let d_obs_nll: f64 = if ruv_mult.is_some() {
-            let mult_p = model.ruv_obs_mult(subject, &theta_p);
-            let nll_p = crate::stats::residual_error::gaussian_obs_nll_scaled(
-                &model.error_spec,
-                &err_keys,
-                &subject.observations,
-                &preds_p,
-                sigma_values,
-                &model.residual_correlations,
-                ruv_scale,
-                frem_ov.as_deref(),
-                mult_p.as_deref(),
-            );
-            (nll_p - nll_base) / delta
-        } else {
-            d_nll_d_f
-                .iter()
-                .zip(preds_p.iter().zip(preds_base.iter()))
-                .map(|(&dl, (&pp, &pb))| dl * (pp - pb) / delta)
-                .sum()
-        };
-        grad[i] = if theta_packs_log_mask[i] {
-            theta[i] * d_obs_nll
-        } else {
-            d_obs_nll
-        };
-    }
-
-    // Sigma gradient: analytical.
-    // d(obs_nll)/d(log_sigma_k) = Σ_j 0.5 * ratio_jk * (1/V_j - resid_j²/V_j²)
-    // where ratio_jk = sigma_k * dV_j/d_sigma_k.
-    for k in 0..n_sigma {
-        let i = n_theta + k;
-        if lower[i] == upper[i] {
-            continue;
-        }
-        let g: f64 = (0..n_obs)
-            .map(|j| {
-                let f = preds_base[j].max(1e-12);
-                let v = variances[j];
-                let resid = residuals[j];
-                // ratio = d(V_j)/d(log sigma_k); zero unless sigma_k enters
-                // obs j's endpoint (so per-CMT each sigma sums only over its
-                // own endpoint's observations). The #484/#1029 magnitude rides
-                // slot k's loading, scaling this derivative by m_k².
-                let ratio = match ruv_mult.as_ref().map(|m| m[j].as_slice()) {
-                    Some(m) => {
-                        model
-                            .error_spec
-                            .dvar_dlogsigma_scaled(err_keys[j], k, f, sigma_values, m)
-                    }
-                    None => model
-                        .error_spec
-                        .dvar_dlogsigma(err_keys[j], k, f, sigma_values),
-                } * obs_var_scale[j];
-                0.5 * ratio * (1.0 / v - resid * resid / (v * v))
-            })
-            .sum();
-        grad[i] = g;
-    }
-
-    (nll_base, grad)
-}
-
 /// Sum of observation log-likelihoods with ETAs held fixed.
 ///
 /// Under M3, censored rows contribute the matching normal-tail likelihood
@@ -1420,7 +898,7 @@ fn obs_nll_subject_grad(
 /// subject the worker handles. With NLopt's central-FD gradient
 /// hitting `obs_nll_sum` `1 + 2·n_dim` times per M-step, this cuts
 /// per-call `Vec<PkParams>` churn to near-zero on TV-cov data.
-fn obs_nll_sum(
+pub(crate) fn obs_nll_sum(
     model: &CompiledModel,
     population: &Population,
     theta: &[f64],
@@ -5446,611 +4924,6 @@ mod tests {
         assert!(unpacked[2] < 0.0);
     }
 
-    /// `obs_nll_subject_grad` summed over subjects must match the reference
-    /// forward-FD of `obs_nll_sum` to within 1e-4 relative tolerance for all
-    /// non-pinned packed parameters (theta + sigma).
-    #[test]
-    fn obs_nll_subject_grad_matches_obs_nll_sum_fd() {
-        use crate::types::{DoseEvent, Population};
-        use std::collections::HashMap;
-
-        let model = analytical_model(GradientMethod::Auto);
-
-        let make_subj = |id: &str, obs: f64| Subject {
-            id: id.into(),
-            doses: vec![DoseEvent::new(0.0, 100.0, 1, 0.0, false, 0.0)],
-            obs_times: vec![1.0, 4.0, 8.0],
-            obs_raw_times: Vec::new(),
-            observations: vec![obs, obs * 0.6, obs * 0.3],
-            obs_cmts: vec![1, 1, 1],
-            covariates: HashMap::new(),
-            dose_covariates: Vec::new(),
-            obs_covariates: Vec::new(),
-            pk_only_times: Vec::new(),
-            pk_only_covariates: Vec::new(),
-            reset_times: Vec::new(),
-            cens: vec![0, 0, 0],
-            occasions: vec![],
-            obs_l2: Vec::new(),
-            dose_occasions: vec![],
-            fremtype: Vec::new(),
-            obs_records: vec![],
-        };
-
-        let population = Population {
-            subjects: vec![
-                make_subj("1", 8.0),
-                make_subj("2", 5.0),
-                make_subj("3", 11.0),
-            ],
-            covariate_names: Vec::new(),
-            dv_column: "DV".into(),
-            input_columns: vec![],
-            exclusions: None,
-            warnings: vec![],
-        };
-
-        let theta = vec![1.5f64, 20.0]; // CL, V
-        let sigma_values = vec![0.2f64]; // proportional
-        let etas: Vec<Vec<f64>> = vec![vec![0.0], vec![0.1], vec![-0.1]];
-        let n_theta = 2;
-        let n_sigma = 1;
-        let n = n_theta + n_sigma;
-
-        // Compute reference gradient via forward-FD of obs_nll_sum.
-        let f0 = obs_nll_sum(&model, &population, &theta, &sigma_values, &etas);
-        let h = 1e-5;
-        let mut ref_grad = vec![0.0f64; n];
-        // Theta perturbations (in natural scale).
-        for i in 0..n_theta {
-            let mut theta_p = theta.clone();
-            theta_p[i] += h;
-            let fp = obs_nll_sum(&model, &population, &theta_p, &sigma_values, &etas);
-            // FD in natural scale; convert to log-packed space (d/d_log = theta * d/d_theta)
-            ref_grad[i] = theta[i] * (fp - f0) / h;
-        }
-        // Sigma perturbation (in natural scale; convert to log-packed).
-        {
-            let mut sigma_p = sigma_values.clone();
-            sigma_p[0] += h;
-            let fp = obs_nll_sum(&model, &population, &theta, &sigma_p, &etas);
-            ref_grad[n_theta] = sigma_values[0] * (fp - f0) / h;
-        }
-
-        // Compute gradient via obs_nll_subject_grad summed over subjects.
-        let mask: Vec<bool> = theta.iter().map(|_| true).collect(); // all log-packed
-        let lo = vec![-1e30f64; n];
-        let hi = vec![1e30f64; n];
-        let mut total_nll = 0.0f64;
-        let mut total_grad = vec![0.0f64; n];
-        let mut scratch = EventPkParams::default();
-        for (i, subject) in population.subjects.iter().enumerate() {
-            let (nll_i, grad_i) = obs_nll_subject_grad(
-                &model,
-                subject,
-                &theta,
-                &sigma_values,
-                &etas[i],
-                &mask,
-                &lo,
-                &hi,
-                n_theta,
-                n_sigma,
-                &mut scratch,
-            );
-            total_nll += nll_i;
-            for (g, gi) in total_grad.iter_mut().zip(grad_i.iter()) {
-                *g += gi;
-            }
-        }
-
-        assert!(
-            (total_nll - f0).abs() < 1e-10,
-            "nll mismatch: {} vs {}",
-            total_nll,
-            f0
-        );
-
-        for j in 0..n {
-            let rel = if ref_grad[j].abs() > 1e-10 {
-                (total_grad[j] - ref_grad[j]).abs() / ref_grad[j].abs()
-            } else {
-                (total_grad[j] - ref_grad[j]).abs()
-            };
-            assert!(
-                rel < 1e-4,
-                "grad[{j}]: obs_nll_subject_grad={:.6e}, ref={:.6e}, rel={:.2e}",
-                total_grad[j],
-                ref_grad[j],
-                rel
-            );
-        }
-    }
-
-    /// IOV M-step gradient (`obs_nll_subject_grad_iov`) must match the forward-FD
-    /// of `obs_nll_subject_into_iov` in log-packed space. This guards the
-    /// analytical gradient that the gradient-based M-step would use — it is not
-    /// exercised by the default BOBYQA M-step (derivative-free), so without this
-    /// direct test the function is untested. Single subject, 2 occasions, κ on CL.
-    #[test]
-    fn obs_nll_subject_grad_iov_matches_fd() {
-        use crate::types::{
-            BloqMethod, CompiledModel, DoseEvent, ErrorModel, ErrorSpec, GradientMethod,
-            ModelParameters, OmegaMatrix, PkModel, PkParams, ScalingSpec, SigmaVector, Subject,
-        };
-        use std::collections::HashMap;
-
-        // Minimal IOV model: CL = TVCL·exp(ETA_CL + KAPPA_CL), V = TVV.
-        let model = CompiledModel {
-            name: "iov_grad_test".into(),
-            pk_model: PkModel::OneCptIv,
-            error_model: ErrorModel::Proportional,
-            error_spec: ErrorSpec::Single(ErrorModel::Proportional),
-            residual_correlations: Vec::new(),
-            pk_param_fn: Box::new(
-                |theta: &[f64], eta: &[f64], _: &HashMap<String, f64>, _t: f64| {
-                    let mut p = PkParams::default();
-                    let kappa = if eta.len() > 1 { eta[1] } else { 0.0 };
-                    p.values[0] = theta[0] * (eta[0] + kappa).exp();
-                    p.values[1] = theta[1];
-                    p
-                },
-            ),
-            n_theta: 2,
-            n_eta: 1,
-            n_epsilon: 1,
-            n_kappa: 1,
-            kappa_names: vec!["KAPPA_CL".into()],
-            theta_names: vec!["TVCL".into(), "TVV".into()],
-            eta_names: vec!["ETA_CL".into()],
-            indiv_param_names: vec!["CL".into(), "V".into()],
-            indiv_param_partials: crate::types::IndivParamPartials::empty(),
-            default_params: ModelParameters {
-                theta: vec![5.0, 50.0],
-                theta_names: vec!["TVCL".into(), "TVV".into()],
-                theta_lower: vec![0.1, 5.0],
-                theta_upper: vec![50.0, 500.0],
-                theta_fixed: vec![false; 2],
-                omega: OmegaMatrix::from_diagonal(&[0.09], vec!["ETA_CL".into()]),
-                omega_fixed: vec![false],
-                sigma: SigmaVector {
-                    values: vec![0.05],
-                    names: vec!["PROP_ERR".into()],
-                },
-                sigma_fixed: vec![false],
-                omega_iov: Some(OmegaMatrix::from_diagonal(&[0.04], vec!["KAPPA_CL".into()])),
-                kappa_fixed: vec![false],
-                mixture: None,
-            },
-            omega_init_as_sd: vec![false],
-            sigma_init_as_sd: vec![false],
-            kappa_init_as_sd: vec![false],
-            kappa_weights: Vec::new(),
-            mu_refs: HashMap::new(),
-            kappa_mu_refs: HashMap::new(),
-            tv_fn: None,
-            pk_indices: vec![0, 1],
-            eta_map: vec![0],
-            pk_idx_f64: vec![0.0, 1.0],
-            sel_flat: vec![1.0, 0.0],
-            ode_spec: None,
-            dose_attr_map: Default::default(),
-            diffusion_theta_start: None,
-            diffusion_state_indices: Vec::new(),
-            bloq_method: BloqMethod::Drop,
-            referenced_covariates: Vec::new(),
-            gradient_method: GradientMethod::Fd,
-            parse_warnings: Vec::new(),
-            has_conditional_eta_params: false,
-            eta_param_info: Vec::new(),
-            theta_transform: Vec::new(),
-            #[cfg(feature = "nn")]
-            covariate_nns: Vec::new(),
-            scaling: ScalingSpec::None,
-            log_transform: false,
-            dv_pre_logged: false,
-            derived_exprs: Vec::new(),
-            output_columns: Vec::new(),
-            #[cfg(feature = "survival")]
-            endpoints: HashMap::new(),
-            frem_config: None,
-            residual_error_eta: None,
-            analytical_init: Vec::new(),
-            analytic_readout: None,
-            ruv_magnitude: None,
-            absorption_ode_equivalent: None,
-            mixture: None,
-        };
-
-        // One subject, 2 occasions (times 1–3 occ 1, 4–6 occ 2), one dose each.
-        let subject = Subject {
-            id: "S1".into(),
-            doses: vec![
-                DoseEvent::new(0.0, 100.0, 1, 0.0, false, 0.0),
-                DoseEvent::new(3.5, 100.0, 1, 0.0, false, 0.0),
-            ],
-            obs_times: vec![1.0, 2.0, 3.0, 4.0, 5.0, 6.0],
-            obs_raw_times: vec![1.0, 2.0, 3.0, 4.0, 5.0, 6.0],
-            observations: vec![36.0, 28.0, 21.0, 34.0, 26.0, 19.0],
-            obs_cmts: vec![1; 6],
-            covariates: HashMap::new(),
-            dose_covariates: Vec::new(),
-            obs_covariates: Vec::new(),
-            pk_only_times: Vec::new(),
-            pk_only_covariates: Vec::new(),
-            reset_times: Vec::new(),
-            cens: vec![0; 6],
-            occasions: vec![1, 1, 1, 2, 2, 2],
-            obs_l2: Vec::new(),
-            dose_occasions: vec![1, 2],
-            fremtype: Vec::new(),
-            obs_records: Vec::new(),
-        };
-
-        let theta = vec![5.0f64, 50.0];
-        let sigma = vec![0.05f64];
-        let eta = vec![0.1f64];
-        let kappas: Vec<Vec<f64>> = vec![vec![0.05], vec![-0.05]]; // one per occasion
-        let n_theta = 2;
-        let n_sigma = 1;
-        let n = n_theta + n_sigma;
-
-        let mut scratch = EventPkParams::default();
-        let (nll, grad) = obs_nll_subject_grad_iov(
-            &model,
-            &subject,
-            &theta,
-            &sigma,
-            &eta,
-            &kappas,
-            &[true, true, true],
-            &[-1e30; 3],
-            &[1e30; 3],
-            n_theta,
-            n_sigma,
-            &mut scratch,
-        );
-
-        // Reference: forward-FD of obs_nll_subject_into_iov in log-packed space.
-        let f0 = obs_nll_subject_into_iov(
-            &model,
-            &subject,
-            &theta,
-            &sigma,
-            &eta,
-            &kappas,
-            &mut scratch,
-        );
-        assert!((nll - f0).abs() < 1e-10, "nll mismatch: {nll} vs {f0}");
-
-        let h = 1e-6;
-        let mut ref_grad = vec![0.0f64; n];
-        for i in 0..n_theta {
-            let mut tp = theta.clone();
-            tp[i] += h;
-            let fp = obs_nll_subject_into_iov(
-                &model,
-                &subject,
-                &tp,
-                &sigma,
-                &eta,
-                &kappas,
-                &mut scratch,
-            );
-            ref_grad[i] = theta[i] * (fp - f0) / h; // d/d_log = theta · d/d_theta
-        }
-        {
-            let mut sp = sigma.clone();
-            sp[0] += h;
-            let fp = obs_nll_subject_into_iov(
-                &model,
-                &subject,
-                &theta,
-                &sp,
-                &eta,
-                &kappas,
-                &mut scratch,
-            );
-            ref_grad[n_theta] = sigma[0] * (fp - f0) / h;
-        }
-
-        for j in 0..n {
-            let rel = if ref_grad[j].abs() > 1e-8 {
-                (grad[j] - ref_grad[j]).abs() / ref_grad[j].abs()
-            } else {
-                (grad[j] - ref_grad[j]).abs()
-            };
-            assert!(
-                rel < 1e-4,
-                "grad[{j}]: analytical={:.6e}, fd={:.6e}, rel={:.2e}",
-                grad[j],
-                ref_grad[j],
-                rel
-            );
-        }
-    }
-
-    /// Per-CMT (multi-endpoint) M-step gradient must match the forward-FD of
-    /// `obs_nll_sum` — the correctness gate for the per-CMT `dvar_df` /
-    /// `dvar_dlogsigma` score terms. Two endpoints with *different* error
-    /// models (proportional PK on CMT=1, additive PD on CMT=2) so a single
-    /// error model would give the wrong Jacobian for one endpoint.
-    #[test]
-    fn obs_nll_subject_grad_per_cmt_matches_fd() {
-        use crate::parser::model_parser::parse_model_string;
-        use crate::types::{DoseEvent, Population};
-        use std::collections::HashMap;
-
-        let model = parse_model_string(
-            r"
-[parameters]
-  theta TVCL(1.0, 0.1, 10.0)
-  theta TVV(10.0, 1.0, 100.0)
-  theta TVKE0(0.5, 0.05, 5.0)
-  omega ETA_CL ~ 0.04
-  sigma PROP_ERR_PK ~ 0.10 (sd)
-  sigma ADD_ERR_PD  ~ 0.50 (sd)
-
-[individual_parameters]
-  CL  = TVCL * exp(ETA_CL)
-  V   = TVV
-  KE0 = TVKE0
-
-[structural_model]
-  ode(states=[central, effect])
-
-[odes]
-  d/dt(central) = -CL/V * central
-  d/dt(effect)  =  KE0 * (central/V - effect)
-
-[scaling]
-  y[CMT=1] = central / V
-  y[CMT=2] = effect
-
-[error_model]
-  CMT=1: DV ~ proportional(PROP_ERR_PK)
-  CMT=2: DV ~ additive(ADD_ERR_PD)
-",
-        )
-        .expect("per-CMT ODE model parses");
-
-        // obs at CMT=1 (PK) and CMT=2 (PD), interleaved.
-        let make_subj = |id: &str, scale: f64| Subject {
-            id: id.into(),
-            doses: vec![DoseEvent::new(0.0, 100.0, 1, 0.0, false, 0.0)],
-            obs_times: vec![1.0, 1.0, 2.0, 2.0, 4.0, 4.0],
-            obs_raw_times: Vec::new(),
-            observations: vec![
-                8.0 * scale,
-                2.0 * scale,
-                6.0 * scale,
-                3.0 * scale,
-                4.0 * scale,
-                3.5 * scale,
-            ],
-            obs_cmts: vec![1, 2, 1, 2, 1, 2],
-            covariates: HashMap::new(),
-            dose_covariates: Vec::new(),
-            obs_covariates: Vec::new(),
-            pk_only_times: Vec::new(),
-            pk_only_covariates: Vec::new(),
-            reset_times: Vec::new(),
-            cens: vec![0; 6],
-            occasions: vec![],
-            obs_l2: Vec::new(),
-            dose_occasions: vec![],
-            fremtype: Vec::new(),
-            obs_records: vec![],
-        };
-        let population = Population {
-            subjects: vec![make_subj("1", 1.0), make_subj("2", 1.1)],
-            covariate_names: Vec::new(),
-            dv_column: "DV".into(),
-            input_columns: vec![],
-            exclusions: None,
-            warnings: vec![],
-        };
-
-        let theta = vec![1.0f64, 10.0, 0.5];
-        let sigma_values = vec![0.10f64, 0.50];
-        let etas: Vec<Vec<f64>> = vec![vec![0.0], vec![0.05]];
-        let n_theta = 3;
-        let n_sigma = 2;
-        let n = n_theta + n_sigma;
-
-        // Reference gradient: forward-FD of obs_nll_sum, in log-packed space.
-        let f0 = obs_nll_sum(&model, &population, &theta, &sigma_values, &etas);
-        let h = 1e-6;
-        let mut ref_grad = vec![0.0f64; n];
-        for i in 0..n_theta {
-            let mut tp = theta.clone();
-            tp[i] += h;
-            let fp = obs_nll_sum(&model, &population, &tp, &sigma_values, &etas);
-            ref_grad[i] = theta[i] * (fp - f0) / h;
-        }
-        for k in 0..n_sigma {
-            let mut sp = sigma_values.clone();
-            sp[k] += h;
-            let fp = obs_nll_sum(&model, &population, &theta, &sp, &etas);
-            ref_grad[n_theta + k] = sigma_values[k] * (fp - f0) / h;
-        }
-
-        // Analytical gradient: sum of per-subject obs_nll_subject_grad.
-        let mask = vec![true; n_theta];
-        let lo = vec![-1e30f64; n];
-        let hi = vec![1e30f64; n];
-        let mut total_nll = 0.0f64;
-        let mut total_grad = vec![0.0f64; n];
-        let mut scratch = EventPkParams::default();
-        for (i, subject) in population.subjects.iter().enumerate() {
-            let (nll_i, grad_i) = obs_nll_subject_grad(
-                &model,
-                subject,
-                &theta,
-                &sigma_values,
-                &etas[i],
-                &mask,
-                &lo,
-                &hi,
-                n_theta,
-                n_sigma,
-                &mut scratch,
-            );
-            total_nll += nll_i;
-            for (g, gi) in total_grad.iter_mut().zip(grad_i.iter()) {
-                *g += gi;
-            }
-        }
-
-        assert!(
-            (total_nll - f0).abs() < 1e-8,
-            "nll mismatch: {total_nll} vs {f0}"
-        );
-        for j in 0..n {
-            let rel = if ref_grad[j].abs() > 1e-8 {
-                (total_grad[j] - ref_grad[j]).abs() / ref_grad[j].abs()
-            } else {
-                (total_grad[j] - ref_grad[j]).abs()
-            };
-            assert!(
-                rel < 1e-3,
-                "per-CMT grad[{j}]: analytical={:.6e}, fd={:.6e}, rel={:.2e}",
-                total_grad[j],
-                ref_grad[j],
-                rel
-            );
-        }
-    }
-
-    /// Dense residual-covariance M-step gradient must match FD of the same
-    /// dense observation NLL. This exercises the `block_sigma` SAEM path, which
-    /// deliberately routes through full FD because the analytic scalar-RUV score
-    /// terms do not apply to off-diagonal R blocks.
-    #[test]
-    fn obs_nll_subject_grad_block_sigma_cross_endpoint_matches_fd() {
-        use crate::parser::model_parser::parse_model_string;
-        use crate::types::{DoseEvent, Population};
-        use std::collections::HashMap;
-
-        let model = parse_model_string(
-            r"
-[parameters]
-  theta TVCL(1.0, 0.1, 10.0)
-  theta TVV(10.0, 1.0, 100.0)
-  omega ETA_CL ~ 0.04
-  block_sigma (PROP_ERR_UNBOUND, PROP_ERR_TOTAL) = [
-    0.04,
-    0.01, 0.09
-  ]
-
-[individual_parameters]
-  CL  = TVCL * exp(ETA_CL)
-  V   = TVV
-
-[structural_model]
-  ode(states=[central])
-
-[odes]
-  d/dt(central) = -CL/V * central
-
-[scaling]
-  y[CMT=1] = 2.0 * central / V
-  y[CMT=2] = central / V
-
-[error_model]
-  CMT=1: DV ~ proportional(PROP_ERR_TOTAL)
-  CMT=2: DV ~ proportional(PROP_ERR_UNBOUND)
-",
-        )
-        .expect("cross-endpoint block_sigma ODE model parses");
-
-        let subject = Subject {
-            id: "1".into(),
-            doses: vec![DoseEvent::new(0.0, 100.0, 1, 0.0, false, 0.0)],
-            obs_times: vec![1.0, 1.0, 2.0, 2.0],
-            obs_raw_times: Vec::new(),
-            observations: vec![17.0, 8.0, 15.0, 7.0],
-            obs_cmts: vec![1, 2, 1, 2],
-            covariates: HashMap::new(),
-            dose_covariates: Vec::new(),
-            obs_covariates: Vec::new(),
-            pk_only_times: Vec::new(),
-            pk_only_covariates: Vec::new(),
-            reset_times: Vec::new(),
-            cens: vec![0; 4],
-            occasions: Vec::new(),
-            obs_l2: Vec::new(),
-            dose_occasions: Vec::new(),
-            fremtype: Vec::new(),
-            obs_records: vec![],
-        };
-        let population = Population {
-            subjects: vec![subject.clone()],
-            covariate_names: Vec::new(),
-            dv_column: "DV".into(),
-            input_columns: vec![],
-            exclusions: None,
-            warnings: vec![],
-        };
-
-        let theta = vec![1.0f64, 10.0];
-        let sigma_values = vec![0.20f64, 0.30];
-        let etas: Vec<Vec<f64>> = vec![vec![0.05]];
-        let n_theta = 2;
-        let n_sigma = 2;
-        let n = n_theta + n_sigma;
-
-        let f0 = obs_nll_sum(&model, &population, &theta, &sigma_values, &etas);
-        let h = 1e-6;
-        let mut ref_grad = vec![0.0f64; n];
-        for i in 0..n_theta {
-            let mut tp = theta.clone();
-            tp[i] += h;
-            let fp = obs_nll_sum(&model, &population, &tp, &sigma_values, &etas);
-            ref_grad[i] = theta[i] * (fp - f0) / h;
-        }
-        for k in 0..n_sigma {
-            let mut sp = sigma_values.clone();
-            sp[k] += h;
-            let fp = obs_nll_sum(&model, &population, &theta, &sp, &etas);
-            ref_grad[n_theta + k] = sigma_values[k] * (fp - f0) / h;
-        }
-
-        let mask = vec![true; n_theta];
-        let lo = vec![-1e30f64; n];
-        let hi = vec![1e30f64; n];
-        let mut scratch = EventPkParams::default();
-        let (nll, grad) = obs_nll_subject_grad(
-            &model,
-            &subject,
-            &theta,
-            &sigma_values,
-            &etas[0],
-            &mask,
-            &lo,
-            &hi,
-            n_theta,
-            n_sigma,
-            &mut scratch,
-        );
-
-        assert!((nll - f0).abs() < 1e-8, "nll mismatch: {nll} vs {f0}");
-        for j in 0..n {
-            let rel = if ref_grad[j].abs() > 1e-8 {
-                (grad[j] - ref_grad[j]).abs() / ref_grad[j].abs()
-            } else {
-                (grad[j] - ref_grad[j]).abs()
-            };
-            assert!(
-                rel < 1e-4,
-                "block_sigma grad[{j}]: fd-path={:.6e}, ref={:.6e}, rel={:.2e}",
-                grad[j],
-                ref_grad[j],
-                rel
-            );
-        }
-    }
-
     // ── IOV kappa MH: rejection restores kappa ─────────────────────────────
 
     /// With `step_scale = 0` the proposal is always identical to the current
@@ -6639,175 +5512,5 @@ mod tests {
             "coordinate with block scale 0 must not move"
         );
         assert_ne!(eta[0], eta0[0], "unclamped coordinate should have moved");
-    }
-
-    // ── #484/#1029: the residual magnitude reaches the SAEM M-step ──────────
-    //
-    // `obs_nll_sum` routes through `likelihood::obs_nll_subject_into`, the same
-    // magnitude-aware data term FOCE/FOCEI score. So asserting the M-step's own
-    // `nll` against it *is* the cross-estimator likelihood-agreement check, and
-    // FD of it pins the θ/σ score terms — including the magnitude's direct-θ
-    // channel, which the prediction chain rule alone would drop.
-
-    /// Shared body: analytic M-step `(nll, grad)` for `model` vs `obs_nll_sum`
-    /// and its forward difference, in the same log-packed space SAEM optimises.
-    fn check_saem_mstep_matches_fd(model: &CompiledModel, theta: &[f64], sigma_values: &[f64]) {
-        use crate::types::{DoseEvent, Population};
-
-        let make_subj = |id: &str, wpse: f64, scale: f64| Subject {
-            id: id.into(),
-            doses: vec![DoseEvent::new(0.0, 100.0, 1, 0.0, false, 0.0)],
-            obs_times: vec![1.0, 4.0, 8.0],
-            observations: vec![8.0 * scale, 6.0 * scale, 4.0 * scale],
-            obs_cmts: vec![1; 3],
-            cens: vec![0; 3],
-            // WPSE varies within the subject: a per-record snapshot, so a
-            // magnitude frozen at the subject's first value would be caught.
-            covariates: [("WPSE".to_string(), wpse)].into_iter().collect(),
-            obs_covariates: vec![
-                [("WPSE".to_string(), wpse)].into_iter().collect(),
-                [("WPSE".to_string(), wpse * 1.5)].into_iter().collect(),
-                [("WPSE".to_string(), wpse * 2.0)].into_iter().collect(),
-            ],
-            ..Default::default()
-        };
-        let population = Population {
-            subjects: vec![make_subj("1", 0.5, 1.0), make_subj("2", 0.8, 1.1)],
-            covariate_names: vec!["WPSE".to_string()],
-            dv_column: "DV".into(),
-            input_columns: vec![],
-            exclusions: None,
-            warnings: vec![],
-        };
-
-        let etas: Vec<Vec<f64>> = vec![vec![0.0], vec![0.05]];
-        let n_theta = theta.len();
-        let n_sigma = sigma_values.len();
-        let n = n_theta + n_sigma;
-
-        let f0 = obs_nll_sum(model, &population, theta, sigma_values, &etas);
-        let h = 1e-6;
-        let mut ref_grad = vec![0.0f64; n];
-        for i in 0..n_theta {
-            let mut tp = theta.to_vec();
-            tp[i] += h;
-            ref_grad[i] =
-                theta[i] * (obs_nll_sum(model, &population, &tp, sigma_values, &etas) - f0) / h;
-        }
-        for k in 0..n_sigma {
-            let mut sp = sigma_values.to_vec();
-            sp[k] += h;
-            ref_grad[n_theta + k] =
-                sigma_values[k] * (obs_nll_sum(model, &population, theta, &sp, &etas) - f0) / h;
-        }
-
-        let mask = vec![true; n_theta];
-        let lo = vec![-1e30f64; n];
-        let hi = vec![1e30f64; n];
-        let mut total_nll = 0.0f64;
-        let mut total_grad = vec![0.0f64; n];
-        let mut scratch = EventPkParams::default();
-        for (i, subject) in population.subjects.iter().enumerate() {
-            let (nll_i, grad_i) = obs_nll_subject_grad(
-                model,
-                subject,
-                theta,
-                sigma_values,
-                &etas[i],
-                &mask,
-                &lo,
-                &hi,
-                n_theta,
-                n_sigma,
-                &mut scratch,
-            );
-            total_nll += nll_i;
-            for (g, gi) in total_grad.iter_mut().zip(grad_i.iter()) {
-                *g += gi;
-            }
-        }
-
-        assert!(
-            (total_nll - f0).abs() < 1e-8,
-            "M-step NLL disagrees with the shared magnitude-aware data term: \
-             {total_nll} vs {f0}"
-        );
-        for j in 0..n {
-            let rel = if ref_grad[j].abs() > 1e-8 {
-                (total_grad[j] - ref_grad[j]).abs() / ref_grad[j].abs()
-            } else {
-                (total_grad[j] - ref_grad[j]).abs()
-            };
-            assert!(
-                rel < 1e-3,
-                "weighted M-step grad[{j}]: analytic={:.6e}, fd={:.6e}, rel={:.2e}",
-                total_grad[j],
-                ref_grad[j],
-                rel
-            );
-        }
-    }
-
-    /// `weight = <covariate>` (#1029): θ-free, so the analytic prediction chain
-    /// rule stays exact once V / ∂V∂f / ∂V∂logσ take their `_scaled` forms.
-    #[test]
-    fn obs_nll_subject_grad_weighted_error_matches_fd() {
-        let model = crate::parser::model_parser::parse_model_string(
-            r"
-[parameters]
-  theta TVCL(1.0, 0.1, 10.0)
-  theta TVV(10.0, 1.0, 100.0)
-  omega ETA_CL ~ 0.04
-  sigma PROP_ERR ~ 0.10 (sd)
-  sigma ADD_ERR  ~ 0.50 (sd)
-[individual_parameters]
-  CL = TVCL * exp(ETA_CL)
-  V  = TVV
-[structural_model]
-  pk one_cpt_iv(cl=CL, v=V)
-[error_model]
-  DV ~ combined(PROP_ERR * (1.0 + 0.5 * WPSE), ADD_ERR) weight = WPSE
-[covariates]
-  WPSE continuous
-",
-        )
-        .expect("weighted model parses");
-        assert!(model.has_custom_ruv_magnitude());
-        assert!(
-            !model.has_theta_dependent_ruv_magnitude(),
-            "a covariate weight must not be flagged θ-dependent"
-        );
-        check_saem_mstep_matches_fd(&model, &[1.0, 10.0], &[0.10, 0.50]);
-    }
-
-    /// θ-*dependent* magnitude (#484): θ now moves the residual variance
-    /// directly as well as through the prediction. The M-step θ gradient must
-    /// carry both channels — the analytic `∂nll/∂f · ∂f/∂θ` chain alone fails
-    /// this test.
-    #[test]
-    fn obs_nll_subject_grad_theta_dependent_magnitude_matches_fd() {
-        let model = crate::parser::model_parser::parse_model_string(
-            r"
-[parameters]
-  theta TVCL(1.0, 0.1, 10.0)
-  theta TVV(10.0, 1.0, 100.0)
-  theta RUV_W(0.30, 0.01, 5.0)
-  omega ETA_CL ~ 0.04
-  sigma PROP_ERR ~ 0.10 (sd)
-  sigma ADD_ERR  ~ 0.50 (sd)
-[individual_parameters]
-  CL = TVCL * exp(ETA_CL)
-  V  = TVV
-[structural_model]
-  pk one_cpt_iv(cl=CL, v=V)
-[error_model]
-  DV ~ combined(PROP_ERR, ADD_ERR * (1.0 + RUV_W * WPSE))
-[covariates]
-  WPSE continuous
-",
-        )
-        .expect("theta-dependent magnitude model parses");
-        assert!(model.has_theta_dependent_ruv_magnitude());
-        check_saem_mstep_matches_fd(&model, &[1.0, 10.0, 0.30], &[0.10, 0.50]);
     }
 }
